Use a clearer initialization syntax.
Specifically, we used the syntax
  Type object = initializer;
in contexts where the initializer is of a different type and the
compiler automatically converted this into
  Type object(initializer);
where the constructor called was just a regular constructor
that did something with the initializer, but specifically did not
copy the object. This is confusing. Use the latter syntax instead.