Break the Manifold interface for performance.
This commit changes the interfaces of

Manifolds::get_default_points_and_weights
Manifold::project_to_manifold
Manifold::get_new_point
Manifold::add_new_points

to use ArrayView instead of std::vector. In addition, the interface of
add_new_points has been changed to populate the ArrayView argument
instead of appending to the end of the array.

This breaks the public interface of Manifold for the sake of improving
performance by about 30%: profiling indicates that, when creating a grid
with a Manifold, we spend about 30% of our time purely calling new and
delete since we must create and destroy so many std::vectors.