Add a note on generating default parameter files.
A since-removed equivalent to parse_input (see commit 8baf33d316b) took,
as an optional argument argument, a boolean that specified whether or
not parse_input should write a default parameter file if one is not
found. This commit summarizes that old documentation for users who want
the old behavior.