Remove using 'deallog' in the linear solvers.
In many cases, all that actually happened was pushing and popping a prefix,
but no actual output. In any case, the level of output was in most cases
not particularly useful to average users (e.g., saying that some restart
happened), and we generally like to have classes and functions that have
no side effects.