Do not set active_fe_indices on non-locally owned cells.
On the other hand, this piece of code tries to do exactly this: after ghost
exchange, we try to set the active_fe_index on a ghost cell to the value
we received from the proper owner of that cell. We work around this by
avoiding cell->set_active_fe_index() on such cells, and manipulate the
underlying data structure directly.

Because these data structures are private, we can no longer do this in a
segregated function in an anonymous namespace, but need to do this in a
function that lives in the 'Implementation' struct.