Allows the CMake option DEAL_II_WITH_CXX11 to be enabled on Windows.
When the CMake option DEAL_II_WITH_CXX11 is enabled on Windows, don`t execute
the test "_check_version("11" "0x")" that checks the compiler flag "-std=c++11".
Since the Visual Studio's compiler doesn`t have a flag to enable/disable c++11,
this test was deactivated. Also, it is made sure that the tests with "CHECK_CXX_SOURCE_COMPILES"
for c++11 are properly executed.