Exit the given subsection in TimerOutput::Scope.
The current implementation doesn't work in a multithreaded context.
Consider the following sequence of events:

Thread 1: start and create Scope scope_1(timer_output, "1")
Thread 2: start and create Scope scope_2(timer_output, "2")
Thread 1: call ~scope_1() and join
Thread 2: call ~scope_2() and join

The current implementation of ~Scope() exits the most recent subsection,
so when ~scope_1() is called we leave subsection "2" and when ~scope_2()
is called we leave subsection "1". We can get around this by always
explicitly exiting the subsection in which we started.