Fix overwriting of unity files in reconfiguration.
One flaw in the unity build system is that it always generates the unity
inclusion files when the build system is (re)configured. This results in a lot
of extra compilation.

This patch fixes the issue by first writing to a temporary file and then
moving it only if it is not equal to the current unity file.