CMake: remove mpi two-pass search
As discussed in https://github.com/dealii/dealii/issues/5510, we must
not assume anything about the internal workings of CMake's FindMPI.cmake
module. In particular, it is not sufficient to set CMAKE_<LANG>_COMPILER
to an MPI-compiler wrapper in order for FindMPI.cmake to actually detect
MPI. So, remove this logic in our FindMPI.cmake module.