Allow copy of Trilinos::MPI::Vector to dealii::Vector on subset of procs
This commit adds a flag that reduces the strictness of the scenario in
which the constructor to dealii::Vector that takes a
Trilinos::MPI::Vector can be called. Previously it was necessary that
this be called on all MPI processors (MPI_COMM_WORLD). By setting the
additional flag to be true, this function need only be called by the
processes (maybe a subset of MPI_COMM_WORLD) that own this vector.