fix MPI type and remove confusing assert
clan-tidy correctly complains about an incorrect MPI type (signed vs
unsigned). While here, remove the assert() that doesn't make much sense:
the subdomain id is not the data to be sent, but the destination and
compilation would fail if it wasn't an int (or convertible to it).
Introduced by @bangerth in 2016.