Fix Vector::operator=(TrilinosWrappers::MPI::Vector)
Data was being copied to the local vector directly from the distributed
vector (as opposed to its localised equivalent). This lead to only a
partial copy of data, and all of the data was offset incorrectly in the
returned vector (giving each process a different view of the localised
vector).