Do not make a variable in step-35 static.
Static variables are only initialized once at the beginning of the run. Here,
we have a vector that is marked as 'static' but this really only works because
we never do mesh refinement in this program and consequently (i) the originally
set size continues to be correct, and (ii) the code below sets every vector
element, rather than add to it.