Enable writing Threads::Thread<T> for types T that are only movable.
Previously, it was not possible to write Threads::new_thread(&foo) or Threads::new_task(&foo)
for functions that return an object whose type is not copyable, but is movable. That's
because in a number of places in the internal machinery of the Threads::Thread and
Threads::Task classes, we didn't account for such types.

This patch fixes this.