Fix issue with AD SymmetricTensor eigenvalue/vector calculations.
For auto-differentiable numbers, when the input tensor was diagonal the
sensitivities of the eigenvalues/vector with respect to one another were
deduced. This means that although the correct results were being
returned, their derivatives were incorrect. This patch attempts to
correct this by perturbing a diagonal input tensor (of AD number type)
and computing an approximation to the eigenvalues/vectors.