Further clarify something about p::d::Triangulation and attached data.
Previously, we said that a particular value is an offset, but this is not
important to a user -- it's just a way to exchange data from one place
to another, but the user does not need to know what this information
corresponds to. Say this, and remove the description of what the data
actually means.