Simplify two functions.
These functions threw an exception in 1d, but this is not
the right place to check that condition: the assertion is currently
thrown because some other class does not implement a function. In
particular, this is about TriaAccessor<0,1>::operator++(), but that
already has the correct assertion that, if ever lifted, would
make the place that currently *also* checks the condition work
automatically.