Fix an issue with CSS on Firefox.
Newer versions of Firefox do not use grid-column-gap or grid-row-gap,
when either value is a percentage instead of a fixed length, when
computing the size of the element. This is not unreasonable; the size of
the container depends on the spacing, which in turn depends on the
spacing, so defining lengths is difficult.

Get around this by just specifying absolute sizes for margins. Since
most of our pictures are 500x500 px this is still about 5% anyway.