Refactor the AD drivers classes.
This commit significantly reworks the implementation of the AD driver
classes. Firstly, all class methods are moved to the associated source
file. The (logically) static class methods have been removed, so all
methods must be called through an instance of a driver class. The
previously implemented logic for existant ADOL-C tapes has been
superceeded by low-level calls to the library itself.

The ADHelper classes are updated to accommodate the changes to the
driver classes.

Fixes #7335
Fixes #7372