DoFTools: restructure set_periodicity_constraints
This commit restructures the function set_periodicity_constraints. Over
the last years (and after fixing about a dozen periodicity related bugs)
this function has gotten out of control. The aim for this restructuring
is to put back some order into the function. This is achieved by:

 - Merging loops and if clauses that were redundant.

 - Minor code cleanup.

 - Switching to a different idiom:

   if(... special case ...)
     {
       // do special case
       continue;
     }

   // next special case

   This saves about 5 (!) levels of indentation and makes the control
   flow significantly more obvious to (a) comment on and (b) understand
   again later.

This commit does not introduce a functional change.