DofTools: Refactor the implementation of make_periodicity_constraints
This commit restructures the actual implementation of
make_periodicity_constraints a bit more. This is in preparation for
fixing an outstanding bug where we might end up with an unstable order
how periodic constraints are being entered on different MPI
subprocesses.

In detail, this commit

 - splits the logic that "collects" periodic constraint entries from the
   one that "enters" periodic constraints into the AffineConstraint
   object.

Due to the fact that this refactoring is hard to review, this commit
does not introduce a functional change. The actual bugfix will be posted
in a separate PR.