Revert "tests: Fix mpi/periodicity_07 test"
Actually, this is not a fix but a bug introduced in mpi/periodicity_07:
The check for consistency in parallel only works if the constraint
object had been closed beforehand.

This means that the order in which we are entering constraints into the
AffineConstraints container actually doesn't matter. This also implies
that my suggested fix of simply rearranging the order in which periodic
boundary conditions and hanging node conditions are applied actually
doesn't fix anything.

This reverts commit b037a42941ef284c20f705496c5b8b50587563b4.